# pkg/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Atelino 后端 API 文档",
        "title": "Atelino API",
        "contact": {},
        "version": "1.0.8"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/user/list": {
            "get": {
                "description": "分页获取所有用户的列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户列表",
                "operationId": "getUserList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数，默认为1",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/atelino_internal_dto.UserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "description": "传入用户的 ID，从数据库中查询指定的用户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户",
                "operationId": "getUserByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "登录请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.TokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出接口，清除用户的会话信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登出",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "未登录或登录状态无效",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "登出失败",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "刷新令牌",
                "operationId": "refreshToken",
                "parameters": [
                    {
                        "description": "刷新令牌请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.TokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "令牌无效或已过期",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "operationId": "register",
                "parameters": [
                    {
                        "description": "注册请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "409": {
                        "description": "该邮箱已注册",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/verify/send": {
            "post": {
                "description": "向指定的邮箱发送验证码用于注册或其他用途。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证"
                ],
                "summary": "发送验证码",
                "operationId": "sendVerificationCode",
                "parameters": [
                    {
                        "description": "邮箱地址",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.SendVerificationCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "429": {
                        "description": "请求过于频繁，请稍后再试",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "发送验证码失败",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "atelino_internal_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.RegisterRequest": {
            "type": "object",
            "required": [
                "code",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "atelino_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.SendVerificationCodeRequest": {
            "type": "object",
            "required": [
                "to"
            ],
            "properties": {
                "to": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}